Fix: Change monitor requset dto

+++ src/main/java/gdg/farm_in_palm/domain/monitor/dto/MonitorInfoResDTO.java
@@ -13,10 +13,10 @@ public class MonitorInfoResDTO {
     private Float humidity;
     private Float groundTemperature;
     private Float groundHumidity;
-    private Integer co2Concentration;
+    private Float co2Concentration;
 
     @Builder
-    public MonitorInfoResDTO(Long monitorId, String monitorName, Float temperature, Float humidity, Float groundTemperature, Float groundHumidity, Integer co2Concentration) {
+    public MonitorInfoResDTO(Long monitorId, String monitorName, Float temperature, Float humidity, Float groundTemperature, Float groundHumidity, Float co2Concentration) {
         this.monitorId = monitorId;
         this.monitorName = monitorName;
         this.temperature = temperature;
